requirements.txt
psutil
uvicorn
gunicorn
fastapi
dotenv
databases
langchain_nomic
langchain_community
langchain-groq
asyncpg
langchain-chroma